backend/requirements.txt
# FastAPI Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Authentication
firebase-admin==6.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# AI/ML Services
groq==0.13.0
pinecone-client==5.0.1
sentence-transformers==3.3.1

# File Processing
pypdf==5.1.0
pytesseract==0.3.13
pillow==11.0.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0

# Security & Middleware
slowapi==0.1.9

# Logging
loguru==0.7.3
